Styling and Script for Nav
- Added a css-file to style navigation (needs to be imported into other css-files, I only added this locally to test out the design).

Left out margins as I believe this should be set by the variables and some form of common styling file for the entire page, to be consistent and align the items ~beautifully~.

Some css-files need editing bc clashing.

Stylin' with having all css-attributes in alphabetical order 8-)

Colors and such are temporary (duh)

- Created nessecsary script-file bc I needed window.location.pathname to mark what page we were on and pug said "no".

Controlled with eslint.

Needs to be added as a script-file to all pages.