Update navigation close/open/underlines
Underlines added to the sublinks, I'm aware that we were supposed to avoid !important, but it was the easiest way to implement the underlines since there are other rules for a-tags.
Also added that the open "menu-dropdowns" closes when the user press on another.
navHam is at the same place on every page (that include navigation.css)