Logo made clickable
The logo is now clickable and takes you back to the homepage. I had to edit the css of the register and planet pages to make it work. A unexpected consequence is that it is now shrinking when the page get's thinner but I tried to prevent it from getting too small while also making sure that the clickable area isn't too big and doesn't match up with what seems visually clickable.